+++ src/Renderer/Renderer.php
@@ -1,8 +1,13 @@
 <?php
+/**
+ * @see       https://github.com/zendframework/zend-expressive-hal for the canonical source repository
+ * @copyright Copyright (c) 2017 Zend Technologies USA Inc. (http://www.zend.com)
+ * @license   https://github.com/zendframework/zend-expressive-hal/blob/master/LICENSE.md New BSD License
+ */
 
-namespace Hal\Renderer;
+namespace Zend\Expressive\Hal\Renderer;
 
-use Hal\HalResource;
+use Zend\Expressive\Hal\HalResource;
 
 interface Renderer
 {