+++ src/Renderer/XmlRenderer.php
@@ -1,11 +1,16 @@
 <?php
+/**
+ * @see       https://github.com/zendframework/zend-expressive-hal for the canonical source repository
+ * @copyright Copyright (c) 2017 Zend Technologies USA Inc. (http://www.zend.com)
+ * @license   https://github.com/zendframework/zend-expressive-hal/blob/master/LICENSE.md New BSD License
+ */
 
-namespace Hal\Renderer;
+namespace Zend\Expressive\Hal\Renderer;
 
 use DOMDocument;
 use DOMNode;
-use Hal\HalResource;
-use Hal\Exception;
+use Zend\Expressive\Hal\HalResource;
+use Zend\Expressive\Hal\Exception;
 
 class XmlRenderer implements Renderer
 {