+++ src/ResourceGenerator.php
@@ -1,6 +1,11 @@
 <?php
+/**
+ * @see       https://github.com/zendframework/zend-expressive-hal for the canonical source repository
+ * @copyright Copyright (c) 2017 Zend Technologies USA Inc. (http://www.zend.com)
+ * @license   https://github.com/zendframework/zend-expressive-hal/blob/master/LICENSE.md New BSD License
+ */
 
-namespace Hal;
+namespace Zend\Expressive\Hal;
 
 use Psr\Container\ContainerInterface;
 use Psr\Http\Message\ServerRequestInterface;