+++ test/HalResourceTest.php
@@ -1,12 +1,17 @@
 <?php
+/**
+ * @see       https://github.com/zendframework/zend-expressive-hal for the canonical source repository
+ * @copyright Copyright (c) 2017 Zend Technologies USA Inc. (http://www.zend.com)
+ * @license   https://github.com/zendframework/zend-expressive-hal/blob/master/LICENSE.md New BSD License
+ */
 
-namespace HalTest;
+namespace ZendTest\Expressive\Hal;
 
-use Hal\HalResource;
-use Hal\Link;
 use InvalidArgumentException;
 use PHPUnit\Framework\TestCase;
 use RuntimeException;
+use Zend\Expressive\Hal\HalResource;
+use Zend\Expressive\Hal\Link;
 
 class HalResourceTest extends TestCase
 {