+++ test/Renderer/TestAsset.php
@@ -1,9 +1,14 @@
 <?php
+/**
+ * @see       https://github.com/zendframework/zend-expressive-hal for the canonical source repository
+ * @copyright Copyright (c) 2017 Zend Technologies USA Inc. (http://www.zend.com)
+ * @license   https://github.com/zendframework/zend-expressive-hal/blob/master/LICENSE.md New BSD License
+ */
 
-namespace HalTest\Renderer;
+namespace ZendTest\Expressive\Hal\Renderer;
 
-use Hal\HalResource;
-use Hal\Link;
+use Zend\Expressive\Hal\HalResource;
+use Zend\Expressive\Hal\Link;
 
 trait TestAsset
 {